requirements.txt
Django==1.6.1
Fabric==1.8.0
Pillow==2.2.2
South==0.8.4
djangorestframework==2.3.10
docutils==0.11
easy-thumbnails==1.4
ecdsa==0.10
paramiko==1.12.0
pycrypto==2.6.1
pytils==0.3
six==1.4.1
wsgiref==0.1.2
